[rename] SignUp.js -> signUp.js, SignUpVerify.js -> signUpVerify.js
