+++ spa/views/signUp.js
